Front thrusters positioned
Ran into another funreal bug. This one caught me out because it said I couldn't save the project and for the cause it was pointing towards the way I set up the thrusters, causing me to think there was an issue in how I coded those up, but nope! The issue was that visual studio needed closing and reopening ffffffff