Car kind of functional?
Car kind of works?? And it does feel fun when it's good, it's challenging to control but it feels like you can master it over time. That said, there's a weird bug causing large jumps from the car that break that mastery and just turn it from hard to frustrating.

My guess is something funky is going on w/ the controls so I'm gonna try and debug that